+++ packages/app/playwright/commons/text.ts
@@ -15,6 +15,14 @@ export async function hasText(
   return textFound;
 }
 
+export async function hasNoText(
+  page: Page,
+  text: string | RegExp,
+  position = 0
+) {
+  return await expect(page.getByText(text).nth(position)).rejects.toThrow();
+}
+
 export async function hasAriaLabel(page: Page, value: string) {
   const selector = await page.waitForSelector(`[aria-label="${value}"]`);
   expect(await selector.getAttribute('aria-label')).toBe(value);