+++ packages/app/src/systems/Home/pages/Home/Home.tsx
@@ -1,13 +1,9 @@
 import { cssObj } from '@fuel-ui/css';
 import { Box } from '@fuel-ui/react';
-import { useMemo } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { VITE_FUEL_PROVIDER_URL } from '~/config';
 import { BalanceWidget, useAccounts } from '~/systems/Account';
-import { MemoAssetList } from '~/systems/Asset';
 import { Layout, Pages, scrollable } from '~/systems/Core';
 import { useBalanceVisibility } from '~/systems/Core/hooks/useVisibility';
-import { useNetworks } from '~/systems/Network';
 
 import { BalanceAssets } from '~/systems/Account/components/BalanceAssets/BalanceAssets';
 import { AssetsTitle, HomeActions } from '../../components';
@@ -15,7 +11,6 @@ import { AssetsTitle, HomeActions } from '../../components';
 export function Home() {
   const { visibility, setVisibility } = useBalanceVisibility();
   const { isLoading, account } = useAccounts();
-  const { selectedNetwork } = useNetworks();
   const navigate = useNavigate();
 
   function sendAction() {
@@ -25,10 +20,6 @@ export function Home() {
   const goToReceive = () => {
     navigate(Pages.receive());
   };
-
-  const emptyProps = useMemo(() => {
-    return { showFaucet: selectedNetwork?.url === VITE_FUEL_PROVIDER_URL };
-  }, [selectedNetwork]);
 
   return (
     <Layout title="Home" isHome>
@@ -54,7 +45,6 @@ export function Home() {
               <BalanceAssets
                 balances={account?.balances || []}
                 isLoading={isLoading}
-                emptyProps={emptyProps}
               />
             </Box.Stack>
           </Box.Stack>