+++ packages/app/src/systems/Transaction/components/TxContent/TxContent.tsx
@@ -4,10 +4,11 @@ import {
   Box,
   CardList,
   ContentLoader,
+  Copyable,
+  Icon,
   Text,
   VStack,
 } from '@fuel-ui/react';
-import type { AssetData } from '@fuel-wallet/types';
 import type { BN, TransactionStatus, TransactionSummary } from 'fuels';
 import { type ReactNode, useMemo } from 'react';
 import { useFormContext } from 'react-hook-form';
@@ -25,14 +26,24 @@ import { TxFeeOptions } from '../TxFeeOptions/TxFeeOptions';
 const ErrorHeader = ({ errors }: { errors?: GroupedErrors }) => {
   return (
     <Alert status="error" css={styles.alert} aria-label="Transaction Error">
-      <Alert.Description
-        as="div"
-        css={{
-          wordBreak: 'break-word',
+      <Copyable
+        value={errors ?? ''}
+        aria-label={errors}
+        iconProps={{
+          icon: Icon.is('Copy'),
+          'aria-label': 'Copy Error',
         }}
+        tooltipMessage="Copy Error"
       >
-        {errors}
-      </Alert.Description>
+        <Alert.Description
+          as="div"
+          css={{
+            wordBreak: 'break-word',
+          }}
+        >
+          {errors}
+        </Alert.Description>
+      </Copyable>
     </Alert>
   );
 };
@@ -66,7 +77,7 @@ function TxContentLoader() {
   );
 }
 
-type TxContentInfoProps = {
+export type TxContentInfoProps = {
   footer?: ReactNode;
   tx?: Maybe<TransactionSummary>;
   txStatus?: Maybe<TransactionStatus>;
@@ -74,7 +85,6 @@ type TxContentInfoProps = {
   isLoading?: boolean;
   isConfirm?: boolean;
   errors?: GroupedErrors;
-  providerUrl?: string;
   fees?: {
     baseFee?: BN;
     minGasLimit?: BN;
@@ -91,7 +101,6 @@ function TxContentInfo({
   isLoading,
   isConfirm,
   errors,
-  providerUrl,
   fees,
 }: TxContentInfoProps) {
   const { getValues } = useFormContext<SendFormValues>();
@@ -114,12 +123,7 @@ function TxContentInfo({
     if (isConfirm) return <ConfirmHeader />;
     if (isExecuted)
       return (
-        <TxHeader
-          id={tx?.id}
-          type={tx?.type}
-          status={status || undefined}
-          providerUrl={providerUrl}
-        />
+        <TxHeader id={tx?.id} type={tx?.type} status={status || undefined} />
       );
 
     return <ConfirmHeader />;